Extract PluginDependencyProvider to interface also fix checkstyle and binary issues
